Added evaporation tower JEI, crowbar tag, element changes, more alloys for future update

+++ src/main/java/com/cannolicatfish/rankine/util/elements/CopperElement.java
@@ -24,7 +24,7 @@ public class CopperElement implements ElementInterface {
     }
 
     @Override
-    public int getDamageFromPercent(int x) {
+    public float getDamageFromPercent(int x) {
         return 0;
     }
