+++ src/main/java/com/cannolicatfish/rankine/util/elements/NickelElement.java
@@ -15,7 +15,7 @@ public class NickelElement implements ElementInterface {
     }
 
     @Override
-    public int getDamageFromPercent(int x) {
+    public float getDamageFromPercent(int x) {
         return 0;
     }
